chore(geofence): move geofence implementation into CioLocationGeofence
Lifts the geofence code out of CioLocation into the new CioLocationGeofence
module. Registering LocationModule alone no longer pulls in geofence
machinery; apps that want geofencing add GeofenceModule alongside it.

What moved (git mv, history preserved)
- All Sources/Location/Geofence/** sources and Tests/Location/Geofence/**
  tests -> Sources/LocationGeofence / Tests/LocationGeofence. The geofence
  DI registrations (sourcery + hand-written) move with them, so Location's
  generated graph drops the geofence accessors and LocationGeofence gains them.

What got re-wired
- GeofenceModule.initialize() now owns the geofence wiring that previously
  lived in LocationModuleState: event subscriptions (ProfileIdentified, Reset,
  LocationAcquired), foreground refresh, first-run rearm, pending-flush, and
  GeofenceBootstrap.wireMonitor. LocationModuleState keeps only the location
  pipeline (cached-location sync + the onAppStart re-identify fix, which is
  mode-aware and therefore stays in Location).
- bootstrapForBackgroundDelivery moves from LocationModule to GeofenceModule.
- GeofenceModule reads the SDK's last known location through the public
  Location API (CustomerIO.location.getLastKnownLocation()), resolved lazily
  so registration order between the two modules doesn't matter.

Public surface
- LocationServices gains getLastKnownLocation() (consumed by geofence, useful
  to hosts). AppLifecycleNotifying / RealAppLifecycleNotifying / token are now
  public SDK-internal infra so geofence can reuse the lifecycle observer.

Sample apps, packaging & CI
- SampleAppsCommon depends on the LocationGeofence product and APN-UIKit
  registers GeofenceModule alongside LocationModule.
- CocoaPods-FCM Podfile adds the CustomerIOLocationGeofence pod and its
  AppDelegate registers GeofenceModule (run `pod install` to pick it up).
- CustomerIO.podspec exposes a LocationGeofence subspec; the release Podfile
  rewrite (reusable_build_sample_apps.yml) versions the new pod for tag builds.

Behavior is unchanged for an app registering both modules. Apps registering
LocationModule only now get location tracking without geofencing.
