+++ Apps/APN-UIKit/APN UIKit/AppDelegate.swift
@@ -1,6 +1,7 @@
 import CioDataPipelines
 import CioInternalCommon
 import CioLocation
+import CioLocationGeofence
 import CioMessagingInApp
 import CioMessagingPush
 import CioMessagingPushAPN
@@ -60,6 +61,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
         let locationMode = settings.location?.trackingMode.toCIOMode() ?? .onAppStart
         config.addModule(LocationModule(config: LocationConfig(mode: locationMode)))
+        config.addModule(GeofenceModule())
         CustomerIO.initialize(withConfig: config.build())
 
         // Initialize messaging features after initializing Customer.io SDK
