+++ Apps/CocoaPods-FCM/Podfile
@@ -11,6 +11,7 @@ target 'test cocoapods' do
   pod 'CustomerIOMessagingInApp', :path => sdk_local_path
   pod 'CustomerIOMessagingPushFCM', :path => sdk_local_path
   pod 'CustomerIOLocation', :path => sdk_local_path
+  pod 'CustomerIOLocationGeofence', :path => sdk_local_path
 
   # Adding APN for internal testing, only. Code from this pod is not used at runtime. 
   # We need to confirm that cocoapods customers can compile their apps with this pod installed.
