+++ Apps/CocoaPods-FCM/src/AppDelegate.swift
@@ -2,6 +2,7 @@ import CioDataPipelines
 import CioFirebaseWrapper
 import CioInternalCommon
 import CioLocation
+import CioLocationGeofence
 import CioMessagingInApp
 import CioMessagingPush
 import CioMessagingPushFCM
@@ -53,6 +54,7 @@ class AppDelegate: NSObject, UIApplicationDelegate {
             config.cdnHost(cdnHost)
         }
         config.addModule(LocationModule(config: LocationConfig(mode: .onAppStart)))
+        config.addModule(GeofenceModule())
         CustomerIO.initialize(withConfig: config.build())
 
         // Initialize messaging features after initializing Customer.io SDK
