+++ Apps/Common/Package.swift
@@ -23,7 +23,8 @@ let package = Package(
                 .product(name: "DataPipelines", package: "customerio-ios"),
                 .product(name: "MessagingPushAPN", package: "customerio-ios"),
                 .product(name: "MessagingInApp", package: "customerio-ios"),
-                .product(name: "Location", package: "customerio-ios")
+                .product(name: "Location", package: "customerio-ios"),
+                .product(name: "LocationGeofence", package: "customerio-ios")
             ],
             path: "Source"
         )
