+++ Sources/Location/AppLifecycleNotifying.swift
@@ -1,14 +1,20 @@
 import Foundation
 
 /// Token returned when registering a lifecycle observer. Retain the token; call `remove()` to unregister, or let it deallocate and the registration is removed automatically.
-protocol AppLifecycleObserverToken: AnyObject {
+///
+/// SDK-internal infrastructure: `public` so the geofence module (which depends on Location)
+/// can reuse it. Not intended for app code.
+public protocol AppLifecycleObserverToken: AnyObject {
     /// Unregisters the observer. Safe to call multiple times; only the first call has effect.
     func remove()
 }
 
 /// Abstracts registration for app lifecycle notifications (didBecomeActive, didEnterBackground).
 /// Enables tests to simulate lifecycle events without NotificationCenter or run loop.
-protocol AppLifecycleNotifying: AnyObject {
+///
+/// SDK-internal infrastructure: `public` so the geofence module (which depends on Location)
+/// can reuse it. Not intended for app code.
+public protocol AppLifecycleNotifying: AnyObject {
     /// Registers for UIApplication.didBecomeActiveNotification. The block is invoked when the app becomes active.
     /// - Parameter block: Called when the notification fires (production: on main queue).
     /// - Returns: Token to retain; call `remove()` to unregister or release the token to unregister on dealloc.
