+++ Sources/Location/LocationModule.swift
@@ -1,6 +1,5 @@
 import CioInternalCommon
 import Foundation
-import UIKit
 
 /// Location module that can be registered via `SDKConfigBuilder.addModule(_:)` so Location is initialized during `CustomerIO.initialize(withConfig:)`.
 ///
@@ -28,29 +27,6 @@ public final class LocationModule: CustomerIOModule {
                 "Location",
                 nil
             )
-        }
-    }
-
-    /// Bootstraps geofence cold-wake delivery from the host's `AppDelegate`.
-    ///
-    /// Wrapper SDKs (RN, Flutter) do not run `CustomerIO.initialize` in the cold-wake
-    /// process — there is no JS/Dart runtime. This entry reads all the state it needs
-    /// from persisted stores and wires up region monitoring + direct-HTTP delivery
-    /// without requiring any module's `initialize` to have run in this process.
-    ///
-    /// Safe to call on every launch. When the SDK has been initialized via
-    /// `CustomerIO.initialize(withConfig:)`, the same DI-resolved singletons are
-    /// reused — no double-init, no duplicate monitoring.
-    @MainActor
-    public static func bootstrapForBackgroundDelivery(launchOptions: [UIApplication.LaunchOptionsKey: Any]?) {
-        // Reserved for future cold-wake detection (e.g. `launchOptions[.location]`).
-        _ = launchOptions
-
-        let di = DIGraphShared.shared
-        GeofenceBootstrap.emitDiscoverabilityLogIfNeeded(di: di)
-        Task { await di.geofenceEventTracker.flushPending() }
-        Task { @MainActor in
-            await GeofenceBootstrap.wireMonitor(di: di)
         }
     }
 }
