+++ Sources/Location/RealAppLifecycleNotifying.swift
@@ -4,14 +4,14 @@ import UIKit
 /// Production implementation that registers for app lifecycle notifications via NotificationCenter.
 /// Observers are invoked on the main queue.
 /// Uses `RegistrationToken` so that when the token is released, the observer is automatically removed.
-final class RealAppLifecycleNotifying: AppLifecycleNotifying {
+public final class RealAppLifecycleNotifying: AppLifecycleNotifying {
     private let notificationCenter: NotificationCenter
 
-    init(notificationCenter: NotificationCenter = .default) {
+    public init(notificationCenter: NotificationCenter = .default) {
         self.notificationCenter = notificationCenter
     }
 
-    func addDidBecomeActiveObserver(using block: @escaping () -> Void) -> AppLifecycleObserverToken {
+    public func addDidBecomeActiveObserver(using block: @escaping () -> Void) -> AppLifecycleObserverToken {
         let observer = notificationCenter.addObserver(
             forName: UIApplication.didBecomeActiveNotification,
             object: nil,
@@ -22,7 +22,7 @@ final class RealAppLifecycleNotifying: AppLifecycleNotifying {
         })
     }
 
-    func addDidEnterBackgroundObserver(using block: @escaping () -> Void) -> AppLifecycleObserverToken {
+    public func addDidEnterBackgroundObserver(using block: @escaping () -> Void) -> AppLifecycleObserverToken {
         let observer = notificationCenter.addObserver(
             forName: UIApplication.didEnterBackgroundNotification,
             object: nil,
