+++ Sources/LocationGeofence/Api/GeofenceApiResponse.swift
