+++ Sources/LocationGeofence/GeofenceConstants.swift
