+++ Sources/LocationGeofence/Logger+Geofence.swift
