+++ Sources/LocationGeofence/Model/Geofence+Distance.swift
