+++ Sources/LocationGeofence/Model/Geofence.swift
