+++ Sources/LocationGeofence/Model/GeofenceConfig.swift
