+++ Sources/LocationGeofence/Model/GeofenceState.swift
