+++ Sources/LocationGeofence/Storage/PendingGeofenceMetric.swift
