+++ Sources/LocationGeofence/Storage/PendingGeofenceMetricStore.swift
