+++ Sources/LocationGeofence/autogenerated/AutoDependencyInjection.generated.swift
@@ -55,10 +55,36 @@ extension DIGraphShared {
     func testDependenciesAbleToResolve() -> Int {
         var countDependenciesResolved = 0
 
+        _ = geofenceApiService
+        countDependenciesResolved += 1
+
+        _ = geofenceEventTracker
+        countDependenciesResolved += 1
+
+        _ = geofenceStorage
+        countDependenciesResolved += 1
+
         return countDependenciesResolved
     }
 
     // Handle classes annotated with InjectRegisterShared
+    // GeofenceApiService (custom. property getter provided via extension)
+    var geofenceApiService: GeofenceApiService {
+        getOverriddenInstance() ??
+            customGeofenceApiService
+    }
+
+    // GeofenceEventTracker (custom. property getter provided via extension)
+    var geofenceEventTracker: GeofenceEventTracker {
+        getOverriddenInstance() ??
+            customGeofenceEventTracker
+    }
+
+    // GeofenceStorage (custom. property getter provided via extension)
+    var geofenceStorage: GeofenceStorage {
+        getOverriddenInstance() ??
+            customGeofenceStorage
+    }
 }
 
 // swiftlint:enable all
