+++ Sources/LocationGeofence/autogenerated/AutoMockable.generated.swift
@@ -81,4 +81,252 @@ import CioInternalCommon
 
  */
 
+/**
+ Class to easily create a mocked version of the `GeofenceApiService` class.
+ This class is equipped with functions and properties ready for you to mock!
+
+ Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
+ See the SDK documentation to learn the basics behind using the mock classes in the SDK.
+ */
+class GeofenceApiServiceMock: GeofenceApiService, Mock {
+    /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
+    var mockCalled: Bool = false //
+
+    init() {}
+
+    public func resetMock() {
+        fetchGeofencesCallsCount = 0
+        fetchGeofencesReceivedArguments = nil
+        fetchGeofencesReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
+    }
+
+    // MARK: - fetchGeofences
+
+    /// Number of times the function was called.
+    @Atomic private(set) var fetchGeofencesCallsCount = 0
+    /// `true` if the function was ever called.
+    var fetchGeofencesCalled: Bool {
+        fetchGeofencesCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    @Atomic private(set) var fetchGeofencesReceivedArguments: (latitude: Double, longitude: Double, completion: (Result<GeofenceApiResponse, GeofenceApiError>) -> Void)?
+    /// Arguments from *all* of the times that the function was called.
+    @Atomic private(set) var fetchGeofencesReceivedInvocations: [(latitude: Double, longitude: Double, completion: (Result<GeofenceApiResponse, GeofenceApiError>) -> Void)] = []
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     */
+    var fetchGeofencesClosure: ((Double, Double, @escaping (Result<GeofenceApiResponse, GeofenceApiError>) -> Void) -> Void)?
+
+    /// Mocked function for `fetchGeofences(latitude: Double, longitude: Double, completion: @escaping (Result<GeofenceApiResponse, GeofenceApiError>) -> Void)`. Your opportunity to return a mocked value and check result of mock in test code.
+    func fetchGeofences(latitude: Double, longitude: Double, completion: @escaping (Result<GeofenceApiResponse, GeofenceApiError>) -> Void) {
+        mockCalled = true
+        fetchGeofencesCallsCount += 1
+        fetchGeofencesReceivedArguments = (latitude: latitude, longitude: longitude, completion: completion)
+        fetchGeofencesReceivedInvocations.append((latitude: latitude, longitude: longitude, completion: completion))
+        fetchGeofencesClosure?(latitude, longitude, completion)
+    }
+}
+
+/**
+ Class to easily create a mocked version of the `GeofenceDeliveryTracker` class.
+ This class is equipped with functions and properties ready for you to mock!
+
+ Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
+ See the SDK documentation to learn the basics behind using the mock classes in the SDK.
+ */
+class GeofenceDeliveryTrackerMock: GeofenceDeliveryTracker, Mock {
+    /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
+    var mockCalled: Bool = false //
+
+    init() {}
+
+    public func resetMock() {
+        trackMetricCallsCount = 0
+        trackMetricReceivedArguments = nil
+        trackMetricReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
+    }
+
+    // MARK: - trackMetric
+
+    /// Number of times the function was called.
+    @Atomic private(set) var trackMetricCallsCount = 0
+    /// `true` if the function was ever called.
+    var trackMetricCalled: Bool {
+        trackMetricCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    @Atomic private(set) var trackMetricReceivedArguments: (metric: PendingGeofenceMetric, userId: String, onComplete: (Result<Void, BackgroundDeliveryHttpError>) -> Void)?
+    /// Arguments from *all* of the times that the function was called.
+    @Atomic private(set) var trackMetricReceivedInvocations: [(metric: PendingGeofenceMetric, userId: String, onComplete: (Result<Void, BackgroundDeliveryHttpError>) -> Void)] = []
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     */
+    var trackMetricClosure: ((PendingGeofenceMetric, String, @escaping (Result<Void, BackgroundDeliveryHttpError>) -> Void) -> Void)?
+
+    /// Mocked function for `trackMetric(metric: PendingGeofenceMetric, userId: String, onComplete: @escaping (Result<Void, BackgroundDeliveryHttpError>) -> Void)`. Your opportunity to return a mocked value and check result of mock in test code.
+    func trackMetric(metric: PendingGeofenceMetric, userId: String, onComplete: @escaping (Result<Void, BackgroundDeliveryHttpError>) -> Void) {
+        mockCalled = true
+        trackMetricCallsCount += 1
+        trackMetricReceivedArguments = (metric: metric, userId: userId, onComplete: onComplete)
+        trackMetricReceivedInvocations.append((metric: metric, userId: userId, onComplete: onComplete))
+        trackMetricClosure?(metric, userId, onComplete)
+    }
+}
+
+/**
+ Class to easily create a mocked version of the `GeofenceSyncCoordinator` class.
+ This class is equipped with functions and properties ready for you to mock!
+
+ Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
+ See the SDK documentation to learn the basics behind using the mock classes in the SDK.
+ */
+class GeofenceSyncCoordinatorMock: GeofenceSyncCoordinator, Mock {
+    /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
+    var mockCalled: Bool = false //
+
+    init() {}
+
+    public func resetMock() {
+        refreshCallsCount = 0
+        refreshReceivedArguments = nil
+        refreshReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
+        handleMovementCallsCount = 0
+        handleMovementReceivedArguments = nil
+        handleMovementReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
+        resetCallsCount = 0
+
+        mockCalled = false // do last as resetting properties above can make this true
+        applyCachedRegistrationCallsCount = 0
+        applyCachedRegistrationReceivedArguments = nil
+        applyCachedRegistrationReceivedInvocations = []
+
+        mockCalled = false // do last as resetting properties above can make this true
+    }
+
+    // MARK: - refresh
+
+    /// Number of times the function was called.
+    @Atomic private(set) var refreshCallsCount = 0
+    /// `true` if the function was ever called.
+    var refreshCalled: Bool {
+        refreshCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    @Atomic private(set) var refreshReceivedArguments: (latitude: Double, longitude: Double)?
+    /// Arguments from *all* of the times that the function was called.
+    @Atomic private(set) var refreshReceivedInvocations: [(latitude: Double, longitude: Double)] = []
+    /// Value to return from the mocked function.
+    var refreshReturnValue: Result<Void, GeofenceSyncError>!
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     The closure has first priority to return a value for the mocked function. If the closure returns `nil`,
+     then the mock will attempt to return the value for `refreshReturnValue`
+     */
+    var refreshClosure: ((Double, Double) -> Result<Void, GeofenceSyncError>)?
+
+    /// Mocked function for `refresh(latitude: Double, longitude: Double)`. Your opportunity to return a mocked value and check result of mock in test code.
+    func refresh(latitude: Double, longitude: Double) -> Result<Void, GeofenceSyncError> {
+        mockCalled = true
+        refreshCallsCount += 1
+        refreshReceivedArguments = (latitude: latitude, longitude: longitude)
+        refreshReceivedInvocations.append((latitude: latitude, longitude: longitude))
+        return refreshClosure.map { $0(latitude, longitude) } ?? refreshReturnValue
+    }
+
+    // MARK: - handleMovement
+
+    /// Number of times the function was called.
+    @Atomic private(set) var handleMovementCallsCount = 0
+    /// `true` if the function was ever called.
+    var handleMovementCalled: Bool {
+        handleMovementCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    @Atomic private(set) var handleMovementReceivedArguments: (latitude: Double, longitude: Double)?
+    /// Arguments from *all* of the times that the function was called.
+    @Atomic private(set) var handleMovementReceivedInvocations: [(latitude: Double, longitude: Double)] = []
+    /// Value to return from the mocked function.
+    var handleMovementReturnValue: Result<Void, GeofenceSyncError>!
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     The closure has first priority to return a value for the mocked function. If the closure returns `nil`,
+     then the mock will attempt to return the value for `handleMovementReturnValue`
+     */
+    var handleMovementClosure: ((Double, Double) -> Result<Void, GeofenceSyncError>)?
+
+    /// Mocked function for `handleMovement(latitude: Double, longitude: Double)`. Your opportunity to return a mocked value and check result of mock in test code.
+    func handleMovement(latitude: Double, longitude: Double) -> Result<Void, GeofenceSyncError> {
+        mockCalled = true
+        handleMovementCallsCount += 1
+        handleMovementReceivedArguments = (latitude: latitude, longitude: longitude)
+        handleMovementReceivedInvocations.append((latitude: latitude, longitude: longitude))
+        return handleMovementClosure.map { $0(latitude, longitude) } ?? handleMovementReturnValue
+    }
+
+    // MARK: - reset
+
+    /// Number of times the function was called.
+    @Atomic private(set) var resetCallsCount = 0
+    /// `true` if the function was ever called.
+    var resetCalled: Bool {
+        resetCallsCount > 0
+    }
+
+    /// Value to return from the mocked function.
+    var resetReturnValue: Result<Void, GeofenceSyncError>!
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     The closure has first priority to return a value for the mocked function. If the closure returns `nil`,
+     then the mock will attempt to return the value for `resetReturnValue`
+     */
+    var resetClosure: (() -> Result<Void, GeofenceSyncError>)?
+
+    /// Mocked function for `reset()`. Your opportunity to return a mocked value and check result of mock in test code.
+    func reset() -> Result<Void, GeofenceSyncError> {
+        mockCalled = true
+        resetCallsCount += 1
+        return resetClosure.map { $0() } ?? resetReturnValue
+    }
+
+    // MARK: - applyCachedRegistration
+
+    /// Number of times the function was called.
+    @Atomic private(set) var applyCachedRegistrationCallsCount = 0
+    /// `true` if the function was ever called.
+    var applyCachedRegistrationCalled: Bool {
+        applyCachedRegistrationCallsCount > 0
+    }
+
+    /// The arguments from the *last* time the function was called.
+    @Atomic private(set) var applyCachedRegistrationReceivedArguments: (cachedRegions: [Geofence], anchor: LocationData?, config: GeofenceConfig?, userId: String?)?
+    /// Arguments from *all* of the times that the function was called.
+    @Atomic private(set) var applyCachedRegistrationReceivedInvocations: [(cachedRegions: [Geofence], anchor: LocationData?, config: GeofenceConfig?, userId: String?)] = []
+    /**
+     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
+     */
+    var applyCachedRegistrationClosure: (([Geofence], LocationData?, GeofenceConfig?, String?) -> Void)?
+
+    /// Mocked function for `applyCachedRegistration(cachedRegions: [Geofence], anchor: LocationData?, config: GeofenceConfig?, userId: String?)`. Your opportunity to return a mocked value and check result of mock in test code.
+    @MainActor
+    func applyCachedRegistration(cachedRegions: [Geofence], anchor: LocationData?, config: GeofenceConfig?, userId: String?) {
+        mockCalled = true
+        applyCachedRegistrationCallsCount += 1
+        applyCachedRegistrationReceivedArguments = (cachedRegions: cachedRegions, anchor: anchor, config: config, userId: userId)
+        applyCachedRegistrationReceivedInvocations.append((cachedRegions: cachedRegions, anchor: anchor, config: config, userId: userId))
+        applyCachedRegistrationClosure?(cachedRegions, anchor, config, userId)
+    }
+}
+
 // swiftlint:enable all
