+++ Tests/LocationGeofence/Geofence/Stub/MockGeofenceRegionMonitor.swift
@@ -1,5 +1,5 @@
 @testable import CioInternalCommon
-@testable import CioLocation
+@testable import CioLocationGeofence
 import Foundation
 
 struct MonitoredRegionRecord: Sendable {
