+++ api-docs/CioLocation.api
@@ -1,3 +1,10 @@
+public interface CioLocation.AppLifecycleNotifying {
+	public fun func addDidBecomeActiveObserver(using block: @escaping () -> Unit) -> AppLifecycleObserverToken;
+	public fun func addDidEnterBackgroundObserver(using block: @escaping () -> Unit) -> AppLifecycleObserverToken;
+}
+public interface CioLocation.AppLifecycleObserverToken {
+	public fun func remove();
+}
 public final class CioLocation.LocationConfig {
 	public final val mode: LocationTrackingMode;
 	public fun init(mode: LocationTrackingMode);
@@ -6,11 +13,16 @@ public final class CioLocation.LocationModule {
 	public final val moduleName: String;
 	public fun init(config: LocationConfig);
 	public fun func initialize();
-	public fun func bootstrapForBackgroundDelivery(launchOptions: List<UIApplication.LaunchOptionsKey: Any>?);
 }
 public interface CioLocation.LocationServices {
 	public fun func setLastKnownLocation(_ location: CLLocation);
 	public fun func requestLocationUpdate();
+	public fun func getLastKnownLocation() async -> LocationData?;
 }
 public enum CioLocation.LocationTrackingMode {
+}
+public final class CioLocation.RealAppLifecycleNotifying {
+	public fun init(notificationCenter: NotificationCenter = .default);
+	public fun func addDidBecomeActiveObserver(using block: @escaping () -> Unit) -> AppLifecycleObserverToken;
+	public fun func addDidEnterBackgroundObserver(using block: @escaping () -> Unit) -> AppLifecycleObserverToken;
 }
